+++ TapResearchSDK.framework/Versions/A/Headers/TapResearch.h
@@ -8,6 +8,9 @@
 #import <Foundation/Foundation.h>
 #import <UIKit/UIColor.h>
 
+@class TRPlacement;
+@class TRReward;
+
 typedef NS_OPTIONS(NSInteger, TRPayoutType) {
     TRPayoutEventProfileComcplete=0,
     TRPayoutDidntQualify = 1,
@@ -15,7 +18,7 @@ typedef NS_OPTIONS(NSInteger, TRPayoutType) {
     TRPayoutEventCampaignComplete = 3
 };
 
-@protocol TapResearchDelegate;
+@protocol TapResearchRewardDelegate;
 @protocol TapResearchSurveyDelegate;
 
 /**
@@ -24,153 +27,80 @@ typedef NS_OPTIONS(NSInteger, TRPayoutType) {
 @interface TapResearch : NSObject
 
 /**
- @method: initWithApiToken:delegate:
- @abstract: Initialize TapResearch with your apiToken and callback handler.
- 
- @param: apiToken Your app's unique identifier.
- @param: delegate: The class that will implement the TapResearchDelegate protocol.
+ @method initWithApiToken:delegate:
+ @abstract Initialize TapResearch with your apiToken and callback handler.
+
+ @param apiToken Your app's unique identifier.
+ @param delegate The class that will implement the TapResearchRewardsDelegate protocol.
  */
-+ (void)initWithApiToken:(NSString *)apiToken delegate:(id<TapResearchDelegate>)delegate;
++ (void)initWithApiToken:(NSString *)apiToken delegate:(id<TapResearchRewardDelegate>)delegate;
 
 /**
- @method: isSurveyAvailable
- @abstract: Check if there is a survey available for the user.
- */
-+ (BOOL)isSurveyAvailable;
-
-/**
- @method: isSurveyAvailable
- @abstract: Check if there is a survey available for the user.
- @param: identifier: offer identifier
- */
-
-+ (BOOL)isSurveyAvailableForIdentifier:(NSString *)identifier;
-
-/**
- @method: showSurvey
- @abstract: Takes the user into a survey. Shows an alert modal asking if the user would like to "cancel" or "continue".
- */
-+ (void)showSurvey;
-
-/**
- @method: showSurveyWithDelegate:
- @abstract: Takes the user into a survey. Shows an alert modal asking if the user would like to "cancel" or "continue".
- @param: delegate: notifies your app which button was selected in the alert modal.
- */
-+ (void)showSurveyWithDelegate:(id<TapResearchSurveyDelegate>)delegate;
-
-/**
- 
- */
-+ (void)showSurveyWithIdentifier:(NSString *)identifier delegate:(id<TapResearchSurveyDelegate>)surveyDelegate;
-
-/**
- @method: setUniqueUserIdentifier
- @abstract: Store unique user identifier for re-identification purposes
- @param: userIdentifier: Unique user identifier. Required if you have opted for server to server postback.
+ @method setUniqueUserIdentifier
+ @abstract Store unique user identifier for re-identification purposes
+ @param userIdentifier Unique user identifier. Required if you have opted for server to server postback.
  */
 + (void)setUniqueUserIdentifier:(NSString *)userIdentifier;
 
 /**
- @method: setNavigationBarColor
- @abstract: Set the survey wall navigation bar color
- @param: color: The navigation bar color
+ @method initPlacementWithIdentifier
+ @abstract Initialize the TRPlacemnt
+*/
++ (void)initPlacementWithIdentifier:(NSString *)placementIdentifier placementBlock:(void(^)(TRPlacement *placement))block;
+/**
+ @method setNavigationBarColor
+ @abstract Set the survey wall navigation bar color
+ @param color The navigation bar color
  */
 + (void)setNavigationBarColor:(UIColor *)color;
 
 /**
- @method: setNavigationBarText
- @abstract: Set the survey wall navigation bar title
- @param: text: Navigation bar title
+ @method setNavigationBarText
+ @abstract Set the survey wall navigation bar title
+ @param text Navigation bar title
  */
 + (void)setNavigationBarText:(NSString *)text;
 
 /**
- @method: setNavigationBarTextColor
- @abstract: Set the navigation bar text color
- @param: color: Navigation bar text color
+ @method setNavigationBarTextColor
+ @abstract Set the navigation bar text color
+ @param color Navigation bar text color
  */
 + (void)setNavigationBarTextColor:(UIColor *)color;
 
++ (void)initWithApiToken:(NSString *)apiToken developmentPlatform:(NSString *)developmentPlatform developmentPlatformVersion:(NSString *)developmentPlatformVersion delegate:(id<TapResearchRewardDelegate>)delegate;
+
 @end
 
-@protocol TapResearchDelegate <NSObject>
+@protocol TapResearchRewardDelegate <NSObject>
 
 /**
- @method: tapResearchDidReceiveRewardWithQuantity:transactionIdentifier:currenyName:payoutEvent
- @abstract: Notifies the delegate that a user has earned an in-app reward.
- 
- @param: quantity: Rewards quantity
- @param: transactionIdentifier: Rewards unique transaction identifier
- @param: currencyName: Currency name, depends on the payoutEvent
- @param: payoutEvent: Payout event type
- @param: offerIdentifier: The offer identifier
+ @method tapResearchDidReceiveRewardWithQuantity:transactionIdentifier:currenyName:payoutEvent
+ @abstract Notifies the delegate that a user has earned an in-app reward.
+ @param reward The new reward
  */
-- (void)tapResearchDidReceiveRewardWithQuantity:(NSInteger)quantity transactionIdentifier:(NSString *)transactionIdentifier
-                                   currencyName:(NSString *)currencyName payoutEvent:(NSInteger)payoutEvent offerIdentifier:(NSString *)offerIdentifier;
-
-@optional
-
-/**
- @method: tapResearchSurveyAvailable
- @abstract: Notifies the delegate when a survey is available.
- */
-- (void)tapResearchOnSurveyAvailable;
-
-/**
- @method: tapResearchOnSurveyAvailableWithPlacement
- @abstract: Notifies the delegate when a survey is available.
- 
- @param: placement: Placement identifier
- */
-- (void)tapResearchOnSurveyAvailableWithPlacement:(NSString *)placement;
-
-/**
- @method: tapResearchOnSurveyNotAvailable
- @abstract: Notifies the delegate when a survey is not available.
- */
-- (void)tapResearchOnSurveyNotAvailable;
-
-/**
- @method: tapResearchOnSurveyNotAvailableWithPlacement
- @abstract: Notifies the delegate when a survey is not available.
-
- @param: placement: Placement identifier
-*/
-- (void)tapResearchOnSurveyNotAvailableWithPlacement:(NSString *)placement;
+- (void)tapResearchDidReceiveReward:(TRReward *)reward;
 
 @end
 
 @protocol TapResearchSurveyDelegate <NSObject>
 
 @optional
-/**
- @method: tapResearchSurveyModalOpened
- @abstract: Notifies the delegate when a user opens the survey modal.
- */
-- (void)tapResearchSurveyModalOpened;
 
 /**
- @method: tapResearchSurveyModalOpened
- @abstract: Notifies the delegate when a user opens the survey modal.
- 
- @param: placement: Placement identifier
+ @method tapResearchSurveyModalOpened
+ @abstract Notifies the delegate when a user opens the survey modal.
+ @param placement Placement identifier
  */
-- (void)tapResearchSurveyModalOpenedWithPlacement:(NSString *)placement;
+- (void)tapResearchSurveyWallOpenedWithPlacement:(TRPlacement *)placement;
 
 /**
- @method: tapResearchSurveyModalDismissed
- @abstract: Notifies the delegate when a user dismisses the survey modal.
+ @method tapResearchSurveyModalDismissed
+ @abstract Notifies the delegate when a user dismisses the survey modal.
+ @param placement Placement identifier
  */
-- (void)tapResearchSurveyModalDismissed;
+- (void)tapResearchSurveyWallDismissedWithPlacement:(TRPlacement *)placement;
 
-/**
- @method: tapResearchSurveyModalDismissed
- @abstract: Notifies the delegate when a user dismisses the survey modal.
- 
- @param: placement: Placement identifier
- */
-- (void)tapResearchSurveyModalDismissedWithPlacement:(NSString *)placement;
 
 @end