+++ TapResearchSDK.framework/Versions/A/Headers/TapResearchSDK.h
@@ -6,3 +6,8 @@
 //
 
 #import "TapResearch.h"
+#import "TRPlacement.h"
+#import "TRReward.h"
+#import "TRSerilizationHelper.h"
+
+